requirements.txt
flask
psycopg2
python-dotenv
waitress
arrow
requests